+++ packages/perseus/src/perseus-types.ts
@@ -160,6 +160,7 @@ export interface PerseusWidgetTypes {
     video: VideoWidget;
 
     // Deprecated widgets
+    "lights-puzzle": AutoCorrectWidget;
     sequence: AutoCorrectWidget;
 }
 
@@ -512,6 +513,7 @@ export type LegacyButtonSets = ReadonlyArray<
     | "logarithms"
     | "basic relations"
     | "advanced relations"
+    | "scientific"
 >;
 
 export type PerseusExpressionWidgetOptions = {
@@ -1218,7 +1220,7 @@ export type PerseusNumericInputAnswer = {
     // Translatable Display; A description for why this answer is correct, wrong, or ungraded
     message: string;
     // The expected answer
-    value?: number;
+    value?: number | null;
     // Whether this answer is "correct", "wrong", or "ungraded"
     status: string;
     // The forms available for this answer.  Options: "integer, ""decimal", "proper", "improper", "mixed", or "pi"
@@ -1297,7 +1299,7 @@ export type PerseusPassageRefWidgetOptions = {
     // The reference number
     referenceNumber: number;
     // Short summary of the referenced section. This will be included in parentheses and quotes automatically.
-    summaryText: string;
+    summaryText?: string;
 };
 
 export const plotterPlotTypes = [