+++ packages/perseus/src/util/parse-perseus-json/perseus-parsers/grapher-widget.ts
@@ -11,6 +11,7 @@ import {
     string,
     union,
 } from "../general-purpose-parsers";
+import {discriminatedUnionOn} from "../general-purpose-parsers/discriminated-union";
 
 import {parseWidget} from "./widget";
 
@@ -35,45 +36,53 @@ export const parseGrapherWidget: Parser<GrapherWidget> = parseWidget(
                 "tangent",
             ),
         ),
-        correct: union(
-            object({
-                type: constant("absolute_value"),
-                coords: pairOfPoints,
-            }),
-        )
-            .or(
+        correct: discriminatedUnionOn("type")
+            .withBranch(
+                "absolute_value",
+                object({
+                    type: constant("absolute_value"),
+                    coords: pairOfPoints,
+                }),
+            )
+            .withBranch(
+                "exponential",
                 object({
                     type: constant("exponential"),
                     asymptote: pairOfPoints,
                     coords: pairOfPoints,
                 }),
             )
-            .or(
+            .withBranch(
+                "linear",
                 object({
                     type: constant("linear"),
                     coords: pairOfPoints,
                 }),
             )
-            .or(
+            .withBranch(
+                "logarithm",
                 object({
                     type: constant("logarithm"),
                     asymptote: pairOfPoints,
                     coords: pairOfPoints,
                 }),
             )
-            .or(
+            .withBranch(
+                "quadratic",
                 object({
                     type: constant("quadratic"),
                     coords: pairOfPoints,
                 }),
             )
-            .or(
+            .withBranch(
+                "sinusoid",
                 object({
                     type: constant("sinusoid"),
                     coords: pairOfPoints,
                 }),
             )
-            .or(
+            .withBranch(
+                "tangent",
                 object({
                     type: constant("tangent"),
                     coords: pairOfPoints,