+++ packages/perseus/src/widgets/dropdown/dropdown.tsx
@@ -80,7 +80,6 @@ class Dropdown extends React.Component<Props> implements Widget {
                     <Renderer
                         content={this.props.placeholder}
                         strings={this.context.strings}
-                        inline
                     />
                 }
                 labelAsText={this.props.placeholder}
@@ -93,7 +92,6 @@ class Dropdown extends React.Component<Props> implements Widget {
                         <Renderer
                             content={choice}
                             strings={this.context.strings}
-                            inline
                         />
                     }
                     labelAsText={choice}
@@ -122,6 +120,7 @@ class Dropdown extends React.Component<Props> implements Widget {
                         <SingleSelect
                             id={dropdownId}
                             placeholder=""
+                            className="perseus-dropdown"
                             onChange={(value) =>
                                 this._handleChange(parseInt(value))
                             }